api/requirements.txt
autopep8==1.5.5
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
pycodestyle==2.6.0
pycparser==2.20
pymongo==3.11.3
python-dotenv==0.15.0
PyYAML==5.4.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
